Split of caches to a separate module.
